requirements.txt
# General dependencies
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.1
PyYAML==6.0.1

# Google Cloud Client Libraries
google-cloud-bigquery==3.22.0
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.49.0 # For Vertex AI (Gemini, Embeddings, Vector Search)
google-cloud-pubsub==2.21.0
google-cloud-secret-manager==2.18.0

# Google AI SDK (for Gemini models)
google-generativeai==0.6.0

# UI/Dashboard specific
streamlit==1.35.0
requests
numpy
alpha_vantage
pandas
# Optional: For better debugging/logging
# opencensus-pep8-compliance
# black
# pylint